MatrixGetEntry angepasst
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@8113 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: afcc8ec13f885cb77c55dccc3649823d832e0051

+++ sensor/streamcars/de.uniol.inf.is.odysseus.scars.emep/src/de/uniol/inf/is/odysseus/scars/emep/functions/MatrixGetEntry.java
@@ -16,7 +16,7 @@ public class MatrixGetEntry extends AbstractFunction<Object>{
 
 	@Override
 	public Object getValue() {
-		return ((double[][]) getInputValue(0))[(Integer)getInputValue(1)][(Integer)getInputValue(2)];
+		return ((double[][]) getInputValue(0))[((Double)getInputValue(1)).intValue()][((Double)getInputValue(2)).intValue()];
 	}
 
 	@Override
